Moved `ReactionHelpers` to a usecase.
